requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=5.0.1=pypi_0
asttokens=2.2.1=pyhd8ed1ab_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blinker=1.6.2=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.5.7=h56e8100_0
cachetools=5.3.1=pypi_0
certifi=2023.5.7=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.1.0=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.5.1=py311hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
fonttools=4.40.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.7.0=pyha770c72_0
importlib_metadata=6.7.0=hd8ed1ab_0
ipykernel=6.15.0=pyh025b116_0
ipython=8.14.0=pyh08f2357_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jsonschema=4.17.3=pypi_0
jupyter_client=8.2.0=pyhd8ed1ab_0
jupyter_core=5.3.1=py311h1ea47a8_0
kiwisolver=1.4.4=pypi_0
libffi=3.4.4=hd77b12b_0
libsodium=1.0.18=h8d14728_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
numpy=1.25.0=pypi_0
openssl=3.0.8=h2bbff1b_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.2=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.5.0=pypi_0
pip=23.1.2=py311haa95532_0
platformdirs=3.6.0=pyhd8ed1ab_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
protobuf=4.23.3=pypi_0
psutil=5.9.0=py311h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=12.0.1=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.15.1=pyhd8ed1ab_0
pympler=1.0.1=pypi_0
pyparsing=3.1.0=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.11.3=he1021f5_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.11=2_cp311
pytz=2023.3=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pywin32=305=py311h2bbff1b_0
pyzmq=25.1.0=py311hd77b12b_0
requests=2.31.0=pypi_0
rich=13.4.2=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
setuptools=67.8.0=py311haa95532_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
streamlit=1.23.1=pypi_0
tenacity=8.2.2=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.3.2=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.6.3=hd8ed1ab_0
typing_extensions=4.6.3=pyha770c72_0
tzdata=2023.3=pypi_0
tzlocal=4.3.1=pypi_0
urllib3=2.0.3=pypi_0
validators=0.20.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=3.0.0=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=py311haa95532_0
xz=5.4.2=h8cc25b3_0
zeromq=4.3.4=h0e60522_1
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
